+++ project/backend/src/app.js
@@ -4,24 +4,25 @@ import dotenv from "dotenv";
 import notesRouter from "./routes/notes.js";
 
 dotenv.config();
-
 const app = express();
 const PORT = process.env.PORT || 5000;
 
+// Middleware
 app.use(cors({ origin: "http://localhost:5173", credentials: true }));
 app.use(express.json());
 
-app.get("/", (_, res) => {
+// Routes
+app.use("/api/notes", notesRouter);
+
+app.get("/", (req, res) => {
   res.json({ message: "CollabNote API is running!" });
 });
 
-app.use("/api/notes", notesRouter);
-
 app.use((err, req, res, next) => {
-  console.error(err.stack);
-  res.status(500).json({ error: "Something went wrong!" });
+  console.error(err);
+  res.status(500).json({ error: "Something went wrong" });
 });
 
 app.listen(PORT, () => {
-  console.log(`🚀 Server running at http://localhost:${PORT}`);
+  console.log(`Server is running on http://localhost:${PORT}`);
 });
